Added missing tests for ReflectionNamedType and ReflectionType
