+++ test/unit/Reflection/ReflectionNamedTypeTest.php
@@ -4,13 +4,18 @@ declare(strict_types=1);
 
 namespace Roave\BetterReflectionTest\Reflection;
 
+use Generator;
 use LogicException;
+use PhpParser\Node\Identifier;
 use PHPUnit\Framework\TestCase;
-use Roave\BetterReflection\BetterReflection;
 use Roave\BetterReflection\Reflection\ReflectionEnum;
+use Roave\BetterReflection\Reflection\ReflectionNamedType;
+use Roave\BetterReflection\Reflection\ReflectionParameter;
 use Roave\BetterReflection\Reflector\DefaultReflector;
+use Roave\BetterReflection\Reflector\Reflector;
 use Roave\BetterReflection\SourceLocator\Ast\Locator;
 use Roave\BetterReflection\SourceLocator\Type\StringSourceLocator;
+use Roave\BetterReflectionTest\BetterReflectionSingleton;
 
 use function sprintf;
 
@@ -19,13 +24,98 @@ use function sprintf;
  */
 class ReflectionNamedTypeTest extends TestCase
 {
+    private Reflector $reflector;
+    private ReflectionParameter $owner;
     private Locator $astLocator;
 
     protected function setUp(): void
     {
         parent::setUp();
 
-        $this->astLocator = (new BetterReflection())->astLocator();
+        $this->reflector  = $this->createMock(Reflector::class);
+        $this->owner      = $this->createMock(ReflectionParameter::class);
+        $this->astLocator = BetterReflectionSingleton::instance()->astLocator();
+    }
+
+    public function testCreateFromNode(): void
+    {
+        $typeInfo = $this->createType('string');
+
+        self::assertInstanceOf(ReflectionNamedType::class, $typeInfo);
+        self::assertSame('string', $typeInfo->getName());
+    }
+
+    public function testAllowsNull(): void
+    {
+        $noNullType = $this->createType('string');
+        self::assertFalse($noNullType->allowsNull());
+
+        $allowsNullType = $this->createType('string', true);
+        self::assertTrue($allowsNullType->allowsNull());
+    }
+
+    public function isBuildinProvider(): Generator
+    {
+        yield ['string'];
+        yield ['int'];
+        yield ['array'];
+        yield ['object'];
+        yield ['iterable'];
+        yield ['mixed'];
+        yield ['never'];
+    }
+
+    /**
+     * @dataProvider isBuildinProvider
+     */
+    public function testIsBuiltin(string $type): void
+    {
+        $reflectionType = $this->createType($type);
+
+        self::assertInstanceOf(ReflectionNamedType::class, $reflectionType);
+        self::assertTrue($reflectionType->isBuiltin());
+    }
+
+    public function isNotBuildinProvider(): Generator
+    {
+        yield ['foo'];
+        yield ['\foo'];
+    }
+
+    /**
+     * @dataProvider isNotBuildinProvider
+     */
+    public function testIsNotBuiltin(string $type): void
+    {
+        $reflectionType = $this->createType($type);
+
+        self::assertInstanceOf(ReflectionNamedType::class, $reflectionType);
+        self::assertFalse($reflectionType->isBuiltin());
+    }
+
+    public function testImplicitCastToString(): void
+    {
+        self::assertSame('int', (string) $this->createType('int'));
+        self::assertSame('?int', (string) $this->createType('int', true));
+        self::assertSame('string', (string) $this->createType('string'));
+        self::assertSame('array', (string) $this->createType('array'));
+        self::assertSame('callable', (string) $this->createType('callable'));
+        self::assertSame('bool', (string) $this->createType('bool'));
+        self::assertSame('float', (string) $this->createType('float'));
+        self::assertSame('void', (string) $this->createType('void'));
+        self::assertSame('object', (string) $this->createType('object'));
+        self::assertSame('iterable', (string) $this->createType('iterable'));
+        self::assertSame('mixed', (string) $this->createType('mixed'));
+        self::assertSame('never', (string) $this->createType('never'));
+
+        self::assertSame('Foo\Bar\Baz', (string) $this->createType('Foo\Bar\Baz'));
+        self::assertSame('\Foo\Bar\Baz', (string) $this->createType('\Foo\Bar\Baz'));
+        self::assertSame('?\Foo\Bar\Baz', (string) $this->createType('\Foo\Bar\Baz', true));
+    }
+
+    private function createType(string $type, bool $allowsNull = false): ReflectionNamedType
+    {
+        return new ReflectionNamedType($this->reflector, $this->owner, new Identifier($type), $allowsNull);
     }
 
     public function testGetClassFromPropertyType(): void
